+++ clients/include/testing_gebrd.hpp
@@ -847,8 +847,18 @@ void testing_gebrd(Arguments& argus)
         }
 
         if(argus.timing)
-            ROCSOLVER_BENCH_INFORM(1);
+            rocsolver_bench_inform(inform_invalid_size);
 
+        return;
+    }
+
+    // memory size query is necessary
+    int size_W;
+    hipsolver_gebrd_bufferSize(FORTRAN, handle, m, n, (T*)nullptr, lda, &size_W);
+
+    if(argus.mem_query)
+    {
+        rocsolver_bench_inform(inform_mem_query, size_W);
         return;
     }
 
@@ -869,6 +879,7 @@ void testing_gebrd(Arguments& argus)
         // device_strided_batch_vector<T>   dTauq(size_Q, 1, stQ, bc);
         // device_strided_batch_vector<T>   dTaup(size_P, 1, stP, bc);
         // device_strided_batch_vector<int> dInfo(1, 1, 1, bc);
+        // device_strided_batch_vector<T>   dWork(size_W, 1, size_W, bc);
         // if(size_A)
         //     CHECK_HIP_ERROR(dA.memcheck());
         // if(size_D)
@@ -880,10 +891,6 @@ void testing_gebrd(Arguments& argus)
         // if(size_P)
         //     CHECK_HIP_ERROR(dTaup.memcheck());
         // CHECK_HIP_ERROR(dInfo.memcheck());
-
-        // int size_W;
-        // hipsolver_gebrd_bufferSize(FORTRAN, handle, m, n, dA.data(), lda, &size_W);
-        // device_strided_batch_vector<T> dWork(size_W, 1, size_W, bc);
         // if(size_W)
         //     CHECK_HIP_ERROR(dWork.memcheck());
 
@@ -966,6 +973,7 @@ void testing_gebrd(Arguments& argus)
         device_strided_batch_vector<T>   dTauq(size_Q, 1, stQ, bc);
         device_strided_batch_vector<T>   dTaup(size_P, 1, stP, bc);
         device_strided_batch_vector<int> dInfo(1, 1, 1, bc);
+        device_strided_batch_vector<T>   dWork(size_W, 1, size_W, bc);
         if(size_A)
             CHECK_HIP_ERROR(dA.memcheck());
         if(size_D)
@@ -977,10 +985,6 @@ void testing_gebrd(Arguments& argus)
         if(size_P)
             CHECK_HIP_ERROR(dTaup.memcheck());
         CHECK_HIP_ERROR(dInfo.memcheck());
-
-        int size_W;
-        hipsolver_gebrd_bufferSize(FORTRAN, handle, m, n, dA.data(), lda, &size_W);
-        device_strided_batch_vector<T> dWork(size_W, 1, size_W, bc);
         if(size_W)
             CHECK_HIP_ERROR(dWork.memcheck());