+++ clients/include/testing_syevd_heevd.hpp
@@ -1,5 +1,5 @@
 /* ************************************************************************
- * Copyright 2021 Advanced Micro Devices, Inc.
+ * Copyright 2021-2022 Advanced Micro Devices, Inc.
  * ************************************************************************ */
 
 #pragma once
@@ -570,8 +570,19 @@ void testing_syevd_heevd(Arguments& argus)
         }
 
         if(argus.timing)
-            ROCSOLVER_BENCH_INFORM(1);
+            rocsolver_bench_inform(inform_invalid_size);
 
+        return;
+    }
+
+    // memory size query is necessary
+    int size_W;
+    hipsolver_syevd_heevd_bufferSize(
+        FORTRAN, handle, evect, uplo, n, (T*)nullptr, lda, (S*)nullptr, &size_W);
+
+    if(argus.mem_query)
+    {
+        rocsolver_bench_inform(inform_mem_query, size_W);
         return;
     }
 
@@ -584,25 +595,21 @@ void testing_syevd_heevd(Arguments& argus)
     // device
     device_strided_batch_vector<S>   dD(size_D, 1, stD, bc);
     device_strided_batch_vector<int> dinfo(1, 1, 1, bc);
+    device_strided_batch_vector<T>   dWork(size_W, 1, size_W, bc);
     if(size_D)
         CHECK_HIP_ERROR(dD.memcheck());
     CHECK_HIP_ERROR(dinfo.memcheck());
+    if(size_W)
+        CHECK_HIP_ERROR(dWork.memcheck());
 
     if(BATCHED)
     {
         // // memory allocations
-        // host_batch_vector<T>   hA(size_A, 1, bc);
-        // host_batch_vector<T>   hAres(size_Ares, 1, bc);
-        // device_batch_vector<T> dA(size_A, 1, bc);
+        // host_batch_vector<T>           hA(size_A, 1, bc);
+        // host_batch_vector<T>           hAres(size_Ares, 1, bc);
+        // device_batch_vector<T>         dA(size_A, 1, bc);
         // if(size_A)
         //     CHECK_HIP_ERROR(dA.memcheck());
-
-        // int size_W;
-        // hipsolver_syevd_heevd_bufferSize(
-        //     FORTRAN, handle, evect, uplo, n, dA.data(), lda, dD.data(), &size_W);
-        // device_strided_batch_vector<T> dWork(size_W, 1, size_W, bc);
-        // if(size_W)
-        //     CHECK_HIP_ERROR(dWork.memcheck());
 
         // // check computations
         // if(argus.unit_check || argus.norm_check)
@@ -663,13 +670,6 @@ void testing_syevd_heevd(Arguments& argus)
         device_strided_batch_vector<T> dA(size_A, 1, stA, bc);
         if(size_A)
             CHECK_HIP_ERROR(dA.memcheck());
-
-        int size_W;
-        hipsolver_syevd_heevd_bufferSize(
-            FORTRAN, handle, evect, uplo, n, dA.data(), lda, dD.data(), &size_W);
-        device_strided_batch_vector<T> dWork(size_W, 1, size_W, bc);
-        if(size_W)
-            CHECK_HIP_ERROR(dWork.memcheck());
 
         // check computations
         if(argus.unit_check || argus.norm_check)