+++ clients/include/testing_syevdx_heevdx.hpp
@@ -797,8 +797,32 @@ void testing_syevdx_heevdx(Arguments& argus)
         }
 
         if(argus.timing)
-            ROCSOLVER_BENCH_INFORM(1);
+            rocsolver_bench_inform(inform_invalid_size);
 
+        return;
+    }
+
+    // memory size query is necessary
+    int size_Work;
+    hipsolver_syevdx_heevdx_bufferSize(API,
+                                       handle,
+                                       evect,
+                                       erange,
+                                       uplo,
+                                       n,
+                                       (T*)nullptr,
+                                       lda,
+                                       vl,
+                                       vu,
+                                       il,
+                                       iu,
+                                       (int*)nullptr,
+                                       (S*)nullptr,
+                                       &size_Work);
+
+    if(argus.mem_query)
+    {
+        rocsolver_bench_inform(inform_mem_query, size_Work);
         return;
     }
 
@@ -813,9 +837,12 @@ void testing_syevdx_heevdx(Arguments& argus)
     // device
     device_strided_batch_vector<S>   dW(size_W, 1, stW, bc);
     device_strided_batch_vector<int> dinfo(1, 1, 1, bc);
+    device_strided_batch_vector<T>   dWork(size_Work, 1, size_Work, bc);
     if(size_W)
         CHECK_HIP_ERROR(dW.memcheck());
     CHECK_HIP_ERROR(dinfo.memcheck());
+    if(size_Work)
+        CHECK_HIP_ERROR(dWork.memcheck());
 
     if(BATCHED)
     {
@@ -825,26 +852,6 @@ void testing_syevdx_heevdx(Arguments& argus)
         // device_batch_vector<T> dA(size_A, 1, bc);
         // if(size_A)
         //     CHECK_HIP_ERROR(dA.memcheck());
-
-        // int size_W;
-        // hipsolver_syevdx_heevdx_bufferSize(API,
-        //                                    handle,
-        //                                    evect,
-        //                                    erange,
-        //                                    uplo,
-        //                                    n,
-        //                                    dA.data(),
-        //                                    lda,
-        //                                    vl,
-        //                                    vu,
-        //                                    il,
-        //                                    iu,
-        //                                    hNevRes.data(),
-        //                                    dW.data(),
-        //                                    &size_W);
-        // device_strided_batch_vector<T> dWork(size_W, 1, size_W, bc);
-        // if(size_W)
-        //     CHECK_HIP_ERROR(dWork.memcheck());
 
         // // check computations
         // if(argus.unit_check || argus.norm_check)
@@ -865,7 +872,7 @@ void testing_syevdx_heevdx(Arguments& argus)
         //                                        dW,
         //                                        stW,
         //                                        dWork,
-        //                                        size_W,
+        //                                        size_Work,
         //                                        dinfo,
         //                                        bc,
         //                                        hA,
@@ -897,7 +904,7 @@ void testing_syevdx_heevdx(Arguments& argus)
         //                                           dW,
         //                                           stW,
         //                                           dWork,
-        //                                           size_W,
+        //                                           size_Work,
         //                                           dinfo,
         //                                           bc,
         //                                           hA,
@@ -920,26 +927,6 @@ void testing_syevdx_heevdx(Arguments& argus)
         if(size_A)
             CHECK_HIP_ERROR(dA.memcheck());
 
-        int size_W;
-        hipsolver_syevdx_heevdx_bufferSize(API,
-                                           handle,
-                                           evect,
-                                           erange,
-                                           uplo,
-                                           n,
-                                           dA.data(),
-                                           lda,
-                                           vl,
-                                           vu,
-                                           il,
-                                           iu,
-                                           hNevRes.data(),
-                                           dW.data(),
-                                           &size_W);
-        device_strided_batch_vector<T> dWork(size_W, 1, size_W, bc);
-        if(size_W)
-            CHECK_HIP_ERROR(dWork.memcheck());
-
         // check computations
         if(argus.unit_check || argus.norm_check)
         {
@@ -959,7 +946,7 @@ void testing_syevdx_heevdx(Arguments& argus)
                                            dW,
                                            stW,
                                            dWork,
-                                           size_W,
+                                           size_Work,
                                            dinfo,
                                            bc,
                                            hA,
@@ -991,7 +978,7 @@ void testing_syevdx_heevdx(Arguments& argus)
                                               dW,
                                               stW,
                                               dWork,
-                                              size_W,
+                                              size_Work,
                                               dinfo,
                                               bc,
                                               hA,